+++ QuoridorGame.View.Bot/Logic/RandomSEV.cs
@@ -7,9 +7,9 @@ namespace QuoridorGame.View.Bot.Logic
 {
     class RandomSEV : IStaticEvaluationFunction
     {
-        public float Eval(Game game) {
+        public double Eval(Game game) {
             Random rand = new Random();
-            return (float)rand.NextDouble();
+            return rand.NextDouble();
         }
     }
 }
